+++ core/pom.xml
@@ -72,6 +72,11 @@
             <artifactId>jorbis</artifactId>
             <version>0.0.17</version>
         </dependency>
+        <dependency>
+            <groupId>fr.delthas</groupId>
+            <artifactId>javamp3</artifactId>
+            <version>1.0.1</version>
+        </dependency>
 
         <!-- mDNS -->
         <dependency>
