+++ core/src/main/java/xyz/gianlu/librespot/player/GeneralAudioStream.java
@@ -13,4 +13,12 @@ public interface GeneralAudioStream {
 
     @NotNull
     String getFileIdHex();
+
+    @NotNull
+    Codec codec();
+
+    enum Codec {
+        VORBIS,
+        STANDARD
+    }
 }
